requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
pillow
opencv-python
pyzbar
pytesseract
passlib[bcrypt]
python-jose
slowapi